tests/wpt/tests/css/css-overflow/scrollable-overflow-transform-003.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Overflow: Scrollable Overflow Transform Inline Element</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-overflow-3/#scrollable" />
<meta name="assert" content="Checks that elements with transform contribute to the scrollable overflow on an inline element.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<style>
  .container {
    display: inline-block;
    width: 100px;
    height: 100px;
    overflow: auto;
    background: silver;
    border: solid thick;
  }

  .element {
    width: 50px;
    height: 50px;
    background: lime;
  }
</style>
<body onload="checkLayout('.container')">

  <div class="container" style="top: 0px;" data-expected-scroll-width="250">
    <div style="transform: translateX(200px);" class="element"></div>
  </div>

  <div class="container" style="top: 150px;" data-expected-scroll-height="350">
    <div style="transform: translateY(300px);" class="element"></div>
  </div>

  <div class="container" style="top: 300px;" data-expected-scroll-width="250" data-expected-scroll-height="350">
    <div style="transform: translate(200px, 300px);" class="element"></div>
  </div>

</body>
